+++ src/androidMain/kotlin/ActualLogger.kt
@@ -6,6 +6,7 @@ import android.util.Log
 internal actual val defaultLogging: (level: LogLevel, tag: String, message: Any, throwable: Throwable?) -> Unit = { l, t, m, e ->
     val messageString = m.toString()
     when(l) {
+        LogLevel.TRACE -> Log.d(t, messageString, e)
         LogLevel.DEBUG -> Log.d(t, messageString, e)
         LogLevel.VERBOSE -> Log.v(t, messageString, e)
         LogLevel.INFO -> Log.i(t, messageString, e)