+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.2.5
 
+* Add `trace` logging level
+
 ## 1.2.4
 
 **This update contains migration onto gradle 8+ (for both wrapper and android plugin). Use it with caution**